chore: move date formatting to a `Date` extension and consolidate `UserDefaults` extensions.
