+++ Common/Sources/Common/Date+.swift
@@ -1,5 +1,5 @@
 //
-//  Date+Extension.swift
+//  Date+.swift
 //  TodoMateInfra
 //
 //  Created by agent on 1/14/26.
@@ -8,35 +8,26 @@
 import Foundation
 
 public extension Date {
-  var formattedForMessage: String {
-    let calendar = Calendar.current
-    let formatter = DateFormatter()
-    if calendar.isDateInToday(self) {
-      formatter.dateFormat = "HH:mm:ss"
-    } else {
-      formatter.dateFormat = "yy/MM/dd"
-    }
-    return formatter.string(from: self)
-  }
-
+  /// Format: `yyyy/MM/dd`
   var yearMonthDay: String {
     let formatter = DateFormatter()
     formatter.dateFormat = "yyyy/MM/dd"
     return formatter.string(from: self)
   }
 
+  /// Format: `yyyy M월`
   var yearMonth: String {
     let dateFormatter = DateFormatter()
     dateFormatter.dateFormat = "yyyy M월"
     return dateFormatter.string(from: self)
   }
 
-  func isSameMonth(as date: Date) -> Bool {
-    Calendar.current.isDate(self, equalTo: date, toGranularity: .month)
-  }
-
   func isSameDay(as date: Date) -> Bool {
     Calendar.current.isDate(self, inSameDayAs: date)
+  }
+
+  func isSameMonth(as date: Date) -> Bool {
+    Calendar.current.isDate(self, equalTo: date, toGranularity: .month)
   }
 
   var isToday: Bool {
@@ -54,13 +45,22 @@ public extension Date {
     return calendar.date(byAdding: .second, value: -1, to: nextDay)!
   }
 
-  // X월 1일 00:00:00으로 초기화
+  var dayRange: ClosedRange<Date> {
+    startOfDay ... endOfDay
+  }
+
+  var monthRange: ClosedRange<Date> {
+    startDayOfMonth ... endDayOfMonth
+  }
+
+  /// 해당 월의 첫 번째 날 (1일 00:00:00)
   var startDayOfMonth: Date {
     let calendar = Calendar.current
     let components = calendar.dateComponents([.year, .month], from: self)
     return calendar.date(from: components) ?? self
   }
 
+  /// 해당 월의 마지막 날 (다음 달 1일의 하루 전)
   var endDayOfMonth: Date {
     let calendar = Calendar.current
     let startOfMonth = calendar.date(from: calendar.dateComponents([.year, .month], from: self))!
@@ -69,11 +69,34 @@ public extension Date {
     return endOfMonth
   }
 
-  var dayRange: ClosedRange<Date> {
-    startOfDay ... endOfDay
+  /// 오늘이면 시간(HH:mm:ss), 아니면 날짜(yy/MM/dd)
+  var timeOrDateString: String {
+    let calendar = Calendar.current
+    let formatter = DateFormatter()
+    if calendar.isDateInToday(self) {
+      formatter.dateFormat = "HH:mm:ss"
+    } else {
+      formatter.dateFormat = "yy/MM/dd"
+    }
+    return formatter.string(from: self)
   }
 
-  var monthRange: ClosedRange<Date> {
-    startDayOfMonth ... endDayOfMonth
+  /// 날짜를 헤더용 문자열로 변환
+  /// - 오늘: "오늘"
+  /// - 어제: "어제"
+  /// - 그 외: "M월 d일"
+  var headerFormatted: String {
+    let calendar = Calendar.current
+    let startOfDate = calendar.startOfDay(for: self)
+    let startOfNow = calendar.startOfDay(for: .now)
+
+    if startOfDate == startOfNow {
+      return "오늘"
+    } else if let yesterday = calendar.date(byAdding: .day, value: -1, to: startOfNow),
+              startOfDate == calendar.startOfDay(for: yesterday) {
+      return "어제"
+    } else {
+      return formatted(.dateTime.month().day())
+    }
   }
 }
